+++ backend/src/main/java/ru/bbnshp/StoreApp.java
@@ -1,4 +1,4 @@
-package org.example;
+package ru.bbnshp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
