+++ backend/src/main/java/ru/bbnshp/entities/OrderStatus.java
@@ -1,4 +1,4 @@
-package org.example.entities;
+package ru.bbnshp.entities;
 
 public enum OrderStatus {
     INPROGRESS, DONE
