+++ backend/src/main/java/ru/bbnshp/repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package org.example.repositories;
+package ru.bbnshp.repositories;
 
-import org.example.entities.User;
+import ru.bbnshp.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
